requirements.txt
qrcode==6.1
reportlab==3.5.66
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tqdm==4.60.0
urllib3==1.26.4
Whoosh==2.7.4
wincertstore==0.2
xlrd==2.0.1
XlsxWriter==1.3.8
xlwt==1.3.0
zipp==3.4.1
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-formtools==2.2
django-haystack==3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-reversion==3.0.9
django-widget-tweaks==1.4.8
docx2pdf==0.1.7
docxtpl==0.11.4
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
jieba==0.42.1
Jinja2==2.11.3
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy
odfpy==1.4.1
olefile==0.46
Pillow==8.1.2
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
pyquery==1.4.3